Move Pi0.5 model files under VLA models
