--- python/sglang/multimodal_gen/runtime/models/pi05/__init__.py
@@ -1,5 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-from sglang.multimodal_gen.runtime.models.pi05.modeling_pi05 import Pi05PolicyModel
-
-__all__ = ["Pi05PolicyModel"]
