+++ python/sglang/multimodal_gen/runtime/models/vlas/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from sglang.multimodal_gen.runtime.models.vlas.pi05_policy import Pi05PolicyModel
+
+__all__ = ["Pi05PolicyModel"]
